requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-mathfilters==1.0.0
django-storages==1.13.1
dropbox==11.35.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
ply==3.11
psycopg2==2.9.3
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
urllib3==1.26.12
